ADD confirm PIN prompts/FIXED PIN 4-8 characters

+++ Encrypto/app/src/main/java/com/example/encrypto/ShowAccount.kt
@@ -1,12 +1,16 @@
 package com.example.encrypto
 
+import android.app.AlertDialog
 import android.os.Bundle
+import android.view.LayoutInflater
+import android.widget.Toast
 import com.google.android.material.snackbar.Snackbar
 import androidx.appcompat.app.AppCompatActivity
 import com.example.encrypto.sql.ManageDB
 
 import kotlinx.android.synthetic.main.activity_show_account.*
 import kotlinx.android.synthetic.main.content_show_account.*
+import kotlinx.android.synthetic.main.dialog_confirm_pin.view.*
 
 class ShowAccount : AppCompatActivity() {
 
@@ -21,13 +25,47 @@ class ShowAccount : AppCompatActivity() {
 
         Username.text = ManageDB().GetUsername(this, selection)
 
-        showPass.setOnClickListener{
-            Password.text = ManageDB().GetPassword(this, selection)
+        showPass.setOnClickListener {
+            val mDialogView = LayoutInflater.from(this).inflate(R.layout.dialog_confirm_pin, null)
+            val mBuilder = AlertDialog.Builder(this)
+                .setView(mDialogView)
+                .setTitle("Confirm")
+            val mAlertDialog = mBuilder.show()
+            mDialogView.confirmdialog_buttonconfirm.setOnClickListener {
+                mAlertDialog.dismiss()
+                val pin = mDialogView.confirmdialog_pin.text.toString()
+                if (ManageDB().CheckPin(this, pin)) {
+                    Password.text = ManageDB().GetPassword(this, selection)
+                } else {
+                    Toast.makeText(this, "Incorrect PIN!\nNo action taken", Toast.LENGTH_LONG)
+                        .show()
+                }
+            }
+            mDialogView.confirmdialog_buttoncancel.setOnClickListener {
+                mAlertDialog.dismiss()
+            }
         }
 
-        btnDelete.setOnClickListener{
-            ManageDB().DeleteAccount(this, selection)
-            finish()
+        btnDelete.setOnClickListener {
+            val mDialogView = LayoutInflater.from(this).inflate(R.layout.dialog_confirm_pin, null)
+            val mBuilder = AlertDialog.Builder(this)
+                .setView(mDialogView)
+                .setTitle("Confirm")
+            val mAlertDialog = mBuilder.show()
+            mDialogView.confirmdialog_buttonconfirm.setOnClickListener {
+                mAlertDialog.dismiss()
+                val pin = mDialogView.confirmdialog_pin.text.toString()
+                if (ManageDB().CheckPin(this, pin)) {
+                    ManageDB().DeleteAccount(this, selection)
+                    finish()
+                } else {
+                    Toast.makeText(this, "Incorrect PIN!\nNo action taken", Toast.LENGTH_LONG)
+                        .show()
+                }
+            }
+            mDialogView.confirmdialog_buttoncancel.setOnClickListener {
+                mAlertDialog.dismiss()
+            }
         }
     }
 
@@ -36,3 +74,4 @@ class ShowAccount : AppCompatActivity() {
         return super.onSupportNavigateUp()
     }
 }
+
